requirements.txt
langchain
langchain-community
langchain-openai
serpapi
python-dotenv
beautifulsoup4
requests